Generators moved to their folder with an index
