+++ generator/entities/generators/effect.js
