+++ generator/entities/generators/index.js
@@ -1,0 +1,4 @@
+export { default as Card } from './card.js';
+export { default as Effect } from './effect.js';
+export { default as MapIcon } from './map-icon.js';
+export { default as Sound } from './sound.js';
