+++ generator/entities/generators/map-icon.js
