+++ generator/sections/202-charms.js
@@ -1,7 +1,5 @@
 import { Section } from '../entities/filter/index.js';
-import MapIcon from '../entities/map-icon.js';
-import Effect from '../entities/effect.js';
-import Card from '../entities/card.js';
+import { Card, Effect, MapIcon } from '../entities/generators/index.js';
 import { global, modes } from '../configuration/index.js';
 
 export default ({ modeId }) => {
