+++ generator/sections/206-jewels.js
@@ -1,7 +1,5 @@
 import { Section } from '../entities/filter/index.js';
-import MapIcon from '../entities/map-icon.js';
-import Effect from '../entities/effect.js';
-import Card from '../entities/card.js';
+import { Card, Effect, MapIcon } from '../entities/generators/index.js';
 
 export default () => {
   const section = Section('Jewels');
