+++ generator/sections/300-gold.js
@@ -1,5 +1,5 @@
 import { Section } from '../entities/filter/index.js';
-import Card from '../entities/card.js';
+import { Card } from '../entities/generators/index.js';
 import { global, modes } from '../configuration/index.js';
 
 export default ({ modeId }) => {
