+++ generator/sections/999-debug.js
@@ -1,8 +1,5 @@
 import { Section } from '../entities/filter/index.js';
-import Sound from '../entities/sound.js';
-import MapIcon from '../entities/map-icon.js';
-import Card from '../entities/card.js';
-import Effect from '../entities/effect.js';
+import { Card, Effect, MapIcon, Sound } from '../entities/generators/index.js';
 import { global } from '../configuration/index.js';
 
 export default () => {
